button/requirements.txt
amqp==2.6.0
appnope==0.1.0
asgiref==3.2.7
attrs==19.3.0
backcall==0.2.0
billiard==3.6.3.0
bleach==3.1.5
celery==4.4.6
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.6
django-debug-toolbar==3.2
django-extensions==3.1.0
django-filter==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
entrypoints==0.3
et-xmlfile==1.0.1
future==0.18.2
idna==2.9
importlib-metadata==4.13.0
inflection==0.5.0
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kombu==4.6.11
Markdown==3.2.2
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
openpyxl==3.0.4
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psycopg2==2.8.6
ptyprocess==0.6.0
pydot==1.4.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9

vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
zipp==3.1.0
mysqlclient==2.0.3
django-cors-headers==3.7.0
qrcode==7.2

django-autocomplete-light==3.8.1
django-mathfilters==1.0.0

python-barcode==0.13.1